Split up app, main, and home page
