+++ app/lib/ui/community_meal_planner_forum_app.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+import 'home_screen.dart';
+
+class CommunityMealPlannerForumApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Community of Meal Planners Forums',
+      theme: ThemeData(
+        primarySwatch: Colors.teal,
+        visualDensity: VisualDensity.adaptivePlatformDensity,
+      ),
+      home: HomeScreen(),
+    );
+  }
+}
